requirements.txt
matplotlib==3.5.1; python_version > '2.7'
mediapipe==0.8.9.1; python_version > '2.7'
numpy==1.24.2; python_version > '2.7'
paho-mqtt==1.6.1; python_version > '2.7'
pyttsx3==2.90; python_version > '2.7'
scipy==1.10.1; python_version > '2.7'
srt==3.5.0; python_version > '2.7'
sounddevice==0.4.4; python_version > '2.7'
vosk==0.3.45; python_version > '2.7'
qibullet==1.4.5; python_version > '2.7'

paho-mqtt==1.6.1; python_version == '2.7'
srt==3.5.0; python_version == '2.7'
